+++ api/src/main/java/com/dnd/sbooky/api/docs/spec/AddLikeApiSpec.java
@@ -1,6 +1,6 @@
 package com.dnd.sbooky.api.docs.spec;
 
-import com.dnd.sbooky.api.member.reqeust.AddLikeRequest;
+import com.dnd.sbooky.api.like.reqeust.AddLikeRequest;
 import com.dnd.sbooky.api.support.response.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
